plymouth: Update build, add eselect module
Since starting to actually use dracut, I noticed that the splash was not
displayed until after switch root occurred, causing quite a noticeable
delay in the boot process and presenting regular status messages.

This patch updates the build to install to the system root (in lieu of
correcting the way dracut generates systemd service files for plymouth)
which allows the binaries to found in the initramfs as they are in a
different location to their system counterparts otherwise.

In addition, this commit converts the kernel symlink eselect module for
use with plymouth and adds it to the build.